chore: move save_existing.go changeset out of common [CLD-2109] (#22289)
* chore: move save_existing.go changeset out of common.

* fix: lint errors
